Factor out increasing_allocator
